requirements.txt
aiohttp==3.6.2
alembic==1.3.1
appdirs==1.4.3
asn1crypto==1.2.0
async-timeout==3.0.1
attrs==19.3.0
cached-property==1.5.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
config==0.4.2
configparser==4.0.2
cryptography==2.6.1
dominate==2.4.0
ecdsa==0.14.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
flask-oidc==1.4.0
flask-paginate==0.5.5
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
future==0.18.2
httplib2==0.14.0
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
jwt==0.3.2
keycloak-wrapper==0.2
Mako==1.1.0
MarkupSafe==1.1.1
multidict==4.6.1
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.18
mysqlclient==1.4.6
oauth2client==4.1.3
packaging==19.2
protobuf==3.11.0
py3oauth2==0.4.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.7.1
pyparsing==2.4.5
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.0.1
python-keycloak==0.17.6
PyYAML==5.2
requests==2.21.0
rsa==4.0
six==1.13.0
SQLAlchemy==1.3.11
typing-extensions==3.7.4.1
urllib3==1.24.3
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
yarl==1.4.1
